Added useState and useEffect hooks to dynamically fetch and display volunteer history from the back-end API.
