+++ client/src/screens/VolunteerHistory.js
@@ -1,47 +1,29 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import '../styles/VolunteerHistory.css'; 
-import logo from '../images/volt2.png'
-
-const volunteerHistory = [
-  {
-    volunteerId: '0001',
-    eventName: 'Beach Cleanup',
-    eventDescription: 'Cleaning the beach for community service.',
-    location: 'Miami Beach, FL',
-    requiredSkills: ['Teamwork', 'Physical Work'],
-    urgency: 'High',
-    eventDate: '2024-09-15',
-    status: 'Completed',
-  },
-  {
-    volunteerId: '0002',
-    eventName: 'Tree Planting',
-    eventDescription: 'Planting trees in the local park.',
-    location: 'Central Park, NY',
-    requiredSkills: ['Gardening', 'Environment Care'],
-    urgency: 'Low',
-    eventDate: '2024-10-01',
-    status: 'Pending',
-  },
-  {
-    volunteerId: '0003',
-    eventName: 'Animal Shelter',
-    eventDescription: 'Assisting with animal care and shelter maintenance',
-    location: 'Chatsworth, CA',
-    requiredSkills: ['Animal Care', 'Cleaning', 'Compassion'],
-    urgency: 'Medium',
-    eventDate: '2024-11-07',
-    status: 'Pending',
-  },
-];
+import logo from '../images/volt2.png'; 
 
 const VolunteerHistory = () => {
+  const [volunteerHistory, setVolunteerHistory] = useState([]);
+
+  useEffect(() => {
+    const fetchHistory = async () => {
+      try {
+        const response = await fetch('http://localhost:4000/api/volunteer-history');
+        const data = await response.json();
+        setVolunteerHistory(data);
+      } catch (error) {
+        console.error('Error fetching volunteer history:', error);
+      }
+    };
+    fetchHistory();
+  }, []);
+
   return (
-    <div>
-      <header className='volunteerheader'>
-      <img src={logo} alt="Logo" className="header-logo" />
+    <div className="volunteer-history">
+      <header className="volunteerheader">
+        <img src={logo} alt="Logo" className="logo-header" />
         Volunteer History
-      </header> 
+      </header>
       <main>
         <div>
           <h1>Volunteer Participation History</h1>
