+++ client/src/styles/VolunteerHistory.css
@@ -1,10 +1,10 @@
 /* Header Styling */
 .volunteerheader {
-  background-color: #0b0b0b; /* Black background for the header */
-  color: #e6e5e3; /* Darker text color */
+  background-color: #0b0b0b; 
+  color: #e6e5e3; 
   padding: 20px;
-  text-align: center; /* Center the text in the header */
-  font-size: 24px; /* Larger font for the header */
+  text-align: center;
+  font-size: 24px;
   font-weight: bold;
 }
 
@@ -12,21 +12,18 @@
   position: center;
   margin-top: 20px;
   top: 15px;
-  left: 400px;
+  left: 200px;
   width: 50px;
   height: auto;
 }
+
 .volunteer-history tbody {
   font-family: Arial, sans-serif;
   background-color: #ffffff;
   margin: 0;
 }
 
-/* .volunteer-history main {
-  padding: 20px;
-}
-
-.volunteer-history #volunteer-history {
+.volunteer-history main {
   padding: 20px;
 }
 
@@ -55,4 +52,4 @@ td {
 
 .volunteer-history tbody tr:nth-child(even) {
   background-color: #ffffff;
-} */
+}
