+++ server/server.js
@@ -10,6 +10,46 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
 
+// Dummy data for Volunteer History (to replace DB for now)
+const volunteerHistory = [
+  {
+    volunteerId: '0001',
+    eventName: 'Beach Cleanup',
+    eventDescription: 'Cleaning the beach for community service.',
+    location: 'Miami Beach, FL',
+    requiredSkills: ['Teamwork', 'Physical Work'],
+    urgency: 'High',
+    eventDate: '2024-09-15',
+    status: 'Completed',
+  },
+  {
+    volunteerId: '0002',
+    eventName: 'Tree Planting',
+    eventDescription: 'Planting trees in the local park.',
+    location: 'Central Park, NY',
+    requiredSkills: ['Gardening', 'Environment Care'],
+    urgency: 'Low',
+    eventDate: '2024-10-01',
+    status: 'Pending',
+  },
+  {
+    volunteerId: '0003',
+    eventName: 'Animal Shelter',
+    eventDescription: 'Assisting with animal care and shelter maintenance',
+    location: 'Chatsworth, CA',
+    requiredSkills: ['Animal Care', 'Cleaning', 'Compassion'],
+    urgency: 'Medium',
+    eventDate: '2024-11-07',
+    status: 'Pending',
+  },
+];
+
+// Route to get volunteer history
+app.get('/api/volunteer-history', (req, res) => {
+  res.json(volunteerHistory);
+});
+
+
 //Error handling
 app.use((err, req, res, next) => {
   const errorStatus = err.status || 500;
